app/requirements.txt
amqp==5.0.9
billiard==3.6.4.0
blinker==1.7.0
celery==5.2.3
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
flask==3.0.0
Flask-Celery-Helper==1.1.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.3
MarkupSafe==2.1.3
prompt-toolkit==3.0.26
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2021.3
six==1.16.0
SQLAlchemy==2.0.24
typing-extensions==4.9.0
vine==5.0.0
wcwidth==0.2.5
werkzeug==3.0.1
zipp==3.17.0
Flask-Migrate==4.0.5
Flask-WTF==1.2.1